refactored embedding directory
