+++ server/src/infra/embed/ollama.embed.js
@@ -1,0 +1,22 @@
+import { Ollama } from 'ollama';
+
+// NOTE: Only Jacob's machine can run the Ollama Embedding Model
+export class OllamaEmbeddingModel extends TemplateEmbeddingModel {
+    constructor() { 
+        super(null); 
+        this.key = process.env.MIXEDBREAD_API_KEY;
+        this.url = 'https://api.mixedbread.com/v1/embeddings';
+    }
+    
+    async initialize() { this.client = new Ollama({ host: 'http://localhost:11434' }).embed; }
+    async _embedImplementation(documents) {
+
+        const vectors = await this.client({
+            model: 'nomic-embed-text',
+            input: documents
+        });
+        
+        return vectors;
+    }
+
+};
